Added user authentication

+++ RestAPI/Backend/Controller/goalController.js
@@ -1,10 +1,10 @@
 const asyncHandler = require('express-async-handler')
 const Goal = require('../models/goalModels')
-
+const User = require('../models/userModel');
 
 
 const getGoals = asyncHandler(async(req,res) => {
-    const goals = await Goal.find()
+    const goals = await Goal.find({ user : req.user.id })
 
     res.status(200).json(goals)
 })
@@ -17,7 +17,8 @@ const setGoals = asyncHandler(async(req,res) => {
     }
 
     const goal = await Goal.create({
-        text : req.body.text 
+        text : req.body.text,
+        user : req.user.id
     })
 
     res.status(200).json(goal)
@@ -30,6 +31,20 @@ const putGoals = asyncHandler(async(req,res) => {
     if(!goal){
         res.status(400)
         throw new Error('Goal not Found');
+    }
+
+    const user = await findById(req.user.id);
+
+    //check for user
+    if(!user){
+        res.status(401);
+        throw new Error('User not Found');
+    }
+
+    //make sure the loged in user matches the goal user
+    if(goal.user.toString() !== user.id){
+        res.status(401);
+        throw new Error('User not Found');
     }
 
     const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.body, {
@@ -46,6 +61,20 @@ const deleteGoals = asyncHandler(async(req,res) => {
     if(!goal){
         res.status(400)
         throw new Error('Goal not Found');
+    }
+
+    const user = await User.findById(req.user.id);
+
+    //check for user
+    if(!user){
+        res.status(401);
+        throw new Error('User not Found');
+    }
+
+    //make sure the loged in user matches the goal user
+    if(goal.user.toString() !== user.id){
+        res.status(401);
+        throw new Error('User not Found');
     }
 
     await goal.deleteOne();
